requirements.txt
monai
absl-py==2.3.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
clearml==2.0.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
crc32c==2.7.1
cycler==0.12.1
databricks-sdk==0.55.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
docker==7.1.0
donfig==0.8.1.post1
einops==0.8.1
executing==2.2.0
fastapi==0.115.12
filelock==3.11.0
fire==0.7.0
Flask==3.1.1
fonttools==4.58.1
fsspec==2025.5.1
furl==2.1.4
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
huggingface-hub==0.32.2
idna==3.10
ignite
imagecodecs==2025.3.30
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itk==5.4.3
itk-core==5.4.3
itk-filtering==5.4.3
itk-io==5.4.3
itk-numerics==5.4.3
itk-registration==5.4.3
itk-segmentation==5.4.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json-tricks==3.17.3
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lpips==0.1.4
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nni==3.0
numcodecs==0.16.1
numpy==1.26.4
nvidia-ml-py==12.575.51
onnx==1.18.0
opencv-python==4.11.0.86
openslide-python==1.4.2
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
optuna==4.3.0
orderedmultidict==1.0.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib2==2.3.7.post1
pillow==11.2.1
platformdirs==4.3.8
prettytable==3.16.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyamg==5.2.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydicom==3.0.1
Pygments==2.19.1
PyJWT==2.10.1
pynrrd==1.1.3
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
PythonWebHDFS==0.2.3
pytorch-ignite
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
responses==0.25.7
rpds-py==0.25.1
rsa==4.9.1
schema==0.7.7
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.26
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.1.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
waitress==3.0.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
zarr==3.0.8
zipp==3.22.0
